+++ src/test/scala/io/findify/s3mock/S3MockTest.scala
@@ -1,6 +1,7 @@
 package io.findify.s3mock
 
 import akka.actor.ActorSystem
+import akka.stream.alpakka.s3.auth.BasicCredentials
 import akka.stream.{ActorMaterializer, Materializer}
 import akka.stream.alpakka.s3.scaladsl.S3Client
 import better.files.File
@@ -30,7 +31,7 @@ trait S3MockTest extends FlatSpec with Matchers with BeforeAndAfterAll {
   private val fileBasedS3 = clientFor("localhost", fileBasedPort)
   private val fileBasedServer = new S3Mock(fileBasedPort, new FileProvider(workDir))
   private val fileBasedTransferManager: TransferManager = TransferManagerBuilder.standard().withS3Client(fileBasedS3).build()
-  private val fileBasedAlpakkaClient = S3Client()(fileSystem, fileMat)
+  private val fileBasedAlpakkaClient = S3Client(BasicCredentials("foo", "bar"), "us-east-1")(fileSystem, fileMat)
 
   private val inMemoryPort = 8002
   private val inMemorySystem = ActorSystem.create("testram", configFor("localhost", inMemoryPort))
@@ -38,7 +39,7 @@ trait S3MockTest extends FlatSpec with Matchers with BeforeAndAfterAll {
   private val inMemoryS3 = clientFor("localhost", inMemoryPort)
   private val inMemoryServer = new S3Mock(inMemoryPort, new InMemoryProvider)
   private val inMemoryTransferManager: TransferManager = TransferManagerBuilder.standard().withS3Client(inMemoryS3).build()
-  private val inMemoryBasedAlpakkaClient = S3Client()(inMemorySystem, inMemoryMat)
+  private val inMemoryBasedAlpakkaClient = S3Client(BasicCredentials("foo", "bar"), "us-east-1")(inMemorySystem, inMemoryMat)
 
   case class Fixture(server: S3Mock, client: AmazonS3, tm: TransferManager, name: String, port: Int, alpakka: S3Client, system: ActorSystem, mat: Materializer)
   val fixtures = List(